// front-end/src/api/model/ootdItemDto.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Wardrobe API
 * The Wardrobe App API description
 * OpenAPI spec version: 1.0
 */

export interface OotdItemDto {
  /** MongoDB _id of the recommended item */
  _id: string;
  /** Item name */
  name: string;
  /** Category name */
  category: string;
  /** Item color */
  color: string;
  /** Image URLs of the item */
  images: string[];
}